Move ATIF datamodels to the core
